+++ yash-prompt/src/prompter.rs
@@ -19,7 +19,7 @@
 use async_trait::async_trait;
 use std::cell::RefCell;
 use yash_env::input::{Context, Input, Result};
-use yash_env::variable::{Expansion, PS1, PS2};
+use yash_env::variable::{Expansion, VariableSet, PS1, PS2};
 use yash_env::Env;
 use yash_syntax::source::Location;
 
@@ -60,23 +60,39 @@ where
 }
 
 async fn print_prompt(env: &mut Env, context: &Context) {
-    let location = Location::dummy(""); // TODO context.location;
-
-    // TODO Yash-specific prompt variables
-    let var = if context.is_first_line() { PS1 } else { PS2 };
-    // https://github.com/rust-lang/rust-clippy/issues/13031
-    #[allow(clippy::manual_unwrap_or_default)]
-    let prompt = match env.variables.get(var).map(|v| v.expand(&location)) {
-        Some(Expansion::Scalar(s)) => s.into_owned(),
-        _ => Default::default(),
-    };
+    // Obtain the prompt string
+    let prompt = fetch_posix(&env.variables, context);
 
     // Perform parameter expansion in the prompt string
-    let expanded_prompt = super::expand_posix(env, &prompt, true).await;
+    let expanded_prompt = super::expand_posix(env, &prompt, context.is_first_line()).await;
 
     // Print the prompt to the standard error
     env.system.print_error(&expanded_prompt).await;
 }
+
+/// Fetches the command prompt string from the variable set.
+///
+/// The return value is the raw value taken from the `PS1` or `PS2` variable
+/// in the set. [`Context::is_first_line`] determines which variable is used.
+/// An empty string is returned if the variable is not found.
+///
+/// The returned prompt string should be expanded before being shown to the
+/// user.
+///
+/// This function does not consider yash-specific prompt variables.
+pub fn fetch_posix(variables: &VariableSet, context: &Context) -> String {
+    // TODO context.location;
+    let location = Location::dummy("");
+
+    let var = if context.is_first_line() { PS1 } else { PS2 };
+    // https://github.com/rust-lang/rust-clippy/issues/13031
+    match variables.get(var).map(|v| v.expand(&location)) {
+        Some(Expansion::Scalar(s)) => s.into_owned(),
+        _ => Default::default(),
+    }
+}
+
+// TODO pub fn fetch_ex: yash-specific prompt variables
 
 #[cfg(test)]
 mod tests {
@@ -145,7 +161,7 @@ mod tests {
         let system = Box::new(VirtualSystem::new());
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        define_variable(&mut env, PS1, "my_custom_prompt>");
+        define_variable(&mut env, PS1, "my_custom_prompt !! >");
         let ref_env = RefCell::new(&mut env);
         let mut prompter = Prompter::new(Memory::new(""), &ref_env);
 
@@ -154,7 +170,8 @@ mod tests {
             .now_or_never()
             .unwrap()
             .ok();
-        assert_stderr(&state, |stderr| assert_eq!(stderr, "my_custom_prompt>"));
+        assert_stderr(&state, |stderr| assert_eq!(stderr, "my_custom_prompt ! >"));
+        // Note that "!!" is expanded to "!" in the prompt string.
     }
 
     #[test]
@@ -162,14 +179,15 @@ mod tests {
         let system = Box::new(VirtualSystem::new());
         let state = Rc::clone(&system.state);
         let mut env = Env::with_system(system);
-        define_variable(&mut env, PS2, "continuation_prompt>");
+        define_variable(&mut env, PS2, "continuation ! >");
         let ref_env = RefCell::new(&mut env);
         let mut prompter = Prompter::new(Memory::new(""), &ref_env);
         let mut context = Context::default();
         context.set_is_first_line(false);
 
         prompter.next_line(&context).now_or_never().unwrap().ok();
-        assert_stderr(&state, |stderr| assert_eq!(stderr, "continuation_prompt>"));
+        assert_stderr(&state, |stderr| assert_eq!(stderr, "continuation ! >"));
+        // Note that "!" is not expanded in the prompt string.
     }
 
     #[test]