+++ yash-builtin/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- The read built-in now shows a prompt when reading a continued line.
 - The break and continue built-ins now return `ExitStatus::ERROR` for syntax
   errors and `ExitStatus::FAILURE` for semantic errors. Previously, they always
   returned `ExitStatus::ERROR` for both types of errors, while the documentation