+++ yash-builtin/CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to `yash-builtin` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] - Unreleased
+## [0.3.0] - Unreleased
 
 ### Changed
 
 - External dependency versions:
     - Rust 1.75.0 → 1.77.0
+    - yash-syntax 0.9.0 → 0.10.0
 
 ### Fixed
 
@@ -77,6 +78,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of the `yash-builtin` crate
 
-[0.2.1]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.2.1
+[0.3.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.3.0
 [0.2.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.2.0
 [0.1.0]: https://github.com/magicant/yash-rs/releases/tag/yash-builtin-0.1.0